PB-454: Reorganized utils
