+++ src/modules/map/components/toolbox/ZoomButtons.vue
@@ -1,7 +1,7 @@
 <script setup>
 import { useStore } from 'vuex'
 
-import { useTippyTooltip } from '@/utils/useTippyTooltip'
+import { useTippyTooltip } from '@/utils/composables/useTippyTooltip'
 
 const dispatcher = { dispatcher: 'ZoomButtons.vue' }
