+++ src/modules/menu/components/activeLayers/MenuActiveLayersListItem.vue
@@ -13,7 +13,7 @@ import MenuActiveLayersListItemTimeSelector from '@/modules/menu/components/acti
 import ErrorButton from '@/utils/components/ErrorButton.vue'
 import TextTruncate from '@/utils/components/TextTruncate.vue'
 import ThirdPartyDisclaimer from '@/utils/components/ThirdPartyDisclaimer.vue'
-import { useTippyTooltip } from '@/utils/useTippyTooltip'
+import { useTippyTooltip } from '@/utils/composables/useTippyTooltip'
 
 const dispatcher = { dispatcher: 'MenuActiveLayersListItem.vue' }
